Fixes #1441: Split system_tests_edge_router into two. Introduced a new system_tests_edge_router1
